Create backup process for personal media
This commit includes a systemd service and timer to call the actual bash
backup script, which uses a restic repository on backblaze b2 storage
same as my laptop backup process. The script is executed by a new user
which has access to the media drive via the media group but does not
have sudo access nor permit ssh login. Also similar to my laptop,
credentials to access the b2 storage are stored as environment variables
in a file in ~/.credentials which are sourced in and out during the
backup script. Though the excludes file is currently empty, it may be
useful in the future.

Prior to this commit, I sorted all files on the media drive. Files which
are personal to me and cannot be found elsewhere (such as photos that I
took) are included in this backup, while media that I can simply
download again (such as movies and TV shows) if the media drive is
damaged are in a separate folder.