smb: Allow connections from any Class C address
Technically this commit widens the attack surface beyond what is
strictly necessary in the case where another subnet separate from the
one which needs access has addresses within the Class C range and can
still reach this host. However, practically speaking I have never
attached the host to a network where that is the case, and instead I
have to add a new subnet range to the smb allow line every time I join a
network which has a never-before-seen range. This change should reduce
maintenance time spent when connecting this host to a new network in the
future.